HB-8: refactor
