+++ src/components/ComponentsUI/components/Icon/utils/applyModificators.ts
