+++ src/tests/mocks.tsx
@@ -1,3 +1,10 @@
+import timekeeper from 'timekeeper';
+
+/**
+ * Mock Date
+ */
+timekeeper.freeze(new Date('2020-01-01'));
+
 /**
  * Hide dirt errors
  */
