+++ src/utils/common/generateID.ts
@@ -1,0 +1,3 @@
+import uniqid from 'uniqid';
+
+export const generateID = (): string => uniqid('', `${new Date().getTime()}`);
